Reorganise crates
client subdir for kubernetes_client
holding and proxy similarly.

The client crate is empty; holding has the current code,
and proxy is also empty, ready for use.

The examples are left in the root crate for discoverability.

No re-exporting done at this time.
