+++ client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kubernetes"
+name = "kubernetes_client"
 version = "0.2.0"
 authors = ["Angus Lees <gus@inodes.org>"]
 description = "A Kubernetes client library for Rust"
@@ -13,18 +13,5 @@ license = "Apache-2.0"
 maintenance = {status = "actively-developed,experimental"}
 
 [dependencies]
-kubernetes_client = { path= "client" }
-kubernetes_holding = { path= "holding" }
-kubernetes_proxy = { path= "proxy" }
 
-
-[dev_dependencies]
-
-failure = "0.1.1"
-futures = "0.1.21"
-hyper = "0.12"
-hyper-tls = "0.3"
-log = "0.4"
-pretty_env_logger = "0.2"
-serde_json = "1.0"
-tokio = "0.1.7"
+[dev-dependencies]
