+++ proxy/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "kubernetes_proxy"
+version = "0.2.0"
+authors = ["Angus Lees <gus@inodes.org>"]
+description = "A K8S API proxy server"
+repository = "https://github.com/anguslees/kubernetes-rs"
+readme = "README.md"
+keywords = ["kubernetes"]
+categories = []
+license = "Apache-2.0"
+
+[badges]
+maintenance = {status = "actively-developed,experimental"}
+
+[dependencies]
+kubernetes_holding = {path="../holding"}
+
+[dev-dependencies]
+pretty_env_logger = "0.2"
+serde_test = "1.0"
